#2601 importing metadata from yaml files (sqlite)

+++ WebPortal/Startup.cs
@@ -162,7 +162,7 @@ namespace NRCan.Datahub.Portal
             InitializeDatabase(logger, etlFactory);
             InitializeDatabase(logger, financeFactory);
             InitializeDatabase(logger, pipFactory);
-            InitializeDatabase(logger, metadataFactory);
+            InitializeDatabase(logger, metadataFactory, false, false);
 
             app.UseRequestLocalization(app.ApplicationServices.GetService<IOptions<RequestLocalizationOptions>>().Value);
 
@@ -194,14 +194,15 @@ namespace NRCan.Datahub.Portal
 
         }
 
-        private void InitializeDatabase<T>(ILogger<Startup> logger, IDbContextFactory<T> factory, bool migrate = true) where T:DbContext
+        private void InitializeDatabase<T>(ILogger<Startup> logger, IDbContextFactory<T> factory, bool migrate = true, bool ensureDelete = true) where T : DbContext
         {
             using var context = factory.CreateDbContext();
             try
             {
                 if (Offline)
                 {
-                    context.Database.EnsureDeleted();
+                    if (ensureDelete)
+                        context.Database.EnsureDeleted();
                     context.Database.EnsureCreated();                    
                 }
                 else
